Consolidate type-specific aggregation files using Go generics
Refactor numeric aggregation implementations to eliminate code duplication:
- Replace 20 type-specific files with 7 generic implementations
- Use Go generics with type aliases for backward compatibility
- Maintain 100% API compatibility and test coverage
- Reduce aggregation package from 40 to 27 files (32% reduction)
